[FIX] point_of_sale: Fix related models and data service
Prevent sending empty array to indexedDB to avoid useless writes and
potential errors.

Check variable before calling a the `isEquals` method on it to avoid
TypeError when the variable is undefined.

X-original-commit: cc9a40558309cb2adfe88515f9e93c422ab563bd

+++ addons/point_of_sale/static/src/app/services/data_service.js
@@ -136,8 +136,13 @@ export class PosData extends Reactive {
                 }
             }
 
-            await this.indexedDB.delete(model, remove);
-            await this.indexedDB.create(model, put);
+            if (remove.length) {
+                await this.indexedDB.delete(model, remove);
+            }
+
+            if (put.length) {
+                await this.indexedDB.create(model, put);
+            }
         }
     }
